+++ src/soap/xsd/IAnySimpleType.cpp
@@ -106,7 +106,7 @@ const AxisString& IAnySimpleType::replaceReservedCharacters(AxisString &value)
     /* Find entity reference characters and returns the first any of chars find
      * position
      */ 
-    unsigned int nPos = value.find_first_of (XML_ENTITY_REFERENCE_CHARS);
+    unsigned long nPos = value.find_first_of (XML_ENTITY_REFERENCE_CHARS);
 
     /* Check for position validity */
     if (AxisString::npos == nPos)
@@ -114,7 +114,7 @@ const AxisString& IAnySimpleType::replaceReservedCharacters(AxisString &value)
         return value;
     }
 
-    int nOldIdx = 0;            // Counter value
+    unsigned long nOldIdx = 0;            // Counter value
     while (AxisString::npos != nPos)
     {                         // Get pointered character
         switch (value.at (nPos))
@@ -147,8 +147,8 @@ const AxisString& IAnySimpleType::replaceReservedCharacters(AxisString &value)
         nPos = value.find_first_of (XML_ENTITY_REFERENCE_CHARS, nPos);
     }
 
-    int nDataLen = value.length ();    // Get the length of the field value
-    int nLen = nDataLen - nOldIdx;      // Get remaining number of characters   
+    unsigned long nDataLen = value.length ();    // Get the length of the field value
+    unsigned long nLen = nDataLen - nOldIdx;      // Get remaining number of characters   
     if (nLen > 0)
     {
         m_strReturnVal += value.substr (nOldIdx, nLen); /* Apend the remaining
